Adding float cast to MOD function
If $a or $b are not strings, you can get an error that fmod needs param 1 ($a) to be of type double.
MS Excel does not fall over when you insert an empty string in MOD, so I'm guessing PHPExcel should do the same?

+++ src/PhpSpreadsheet/Calculation/MathTrig.php
@@ -673,8 +673,8 @@ class MathTrig
      */
     public static function MOD($a = 1, $b = 1)
     {
-        $a = Functions::flattenSingleValue($a);
-        $b = Functions::flattenSingleValue($b);
+        $a = (float) Functions::flattenSingleValue($a);
+        $b = (float) Functions::flattenSingleValue($b);
 
         if ($b == 0.0) {
             return Functions::DIV0();
